Fixed API issue
